Code samengevoegd, erros opgelost

+++ Core/Inc/IOL.h
@@ -14,6 +14,9 @@ void screen(void);
 #include "FL.h"
 
 #include "bitmaps.h"
+#include "arial_fonts.h"
+#include "consolas_fonts.h"
+
 //#include "screen.c"
 //#include "images.h"
 
@@ -28,9 +31,9 @@ void screen(void);
 typedef unsigned char  byte;
 typedef unsigned short word;
 
-int IO_draw_rectangle(int x_lup, int y_lup, int width, int height, int color, int filled);
+int IO_draw_rectangle(uint16_t x_lup, uint16_t y_lup, int width, int height, int color, int filled);
 int IO_clearscreen(int color);
-int IO_draw_line(int x1, int y1, int x2, int y2, byte color, int thickness);
+int IO_draw_line(uint16_t x1, uint16_t y1, uint16_t x2, uint16_t y2, byte color, int thickness);
 int IO_draw_circle(int xc, int yc, int radius, byte color);
 int drawCircle(int xc, int yc, int x, int y, byte color);
 int IO_draw_figure(uint16_t x1,uint16_t y1,uint16_t x2,uint16_t y2,uint16_t x3,uint16_t y3,uint16_t x4,uint16_t y4,uint16_t x5,uint16_t y5, byte color);
